Feature .Graduate work. Package, utility, get event, get user, postman

+++ ewm-service/src/main/java/ru/practicum/ewmservice/event/dto/EventFullPublicDto.java
@@ -54,7 +54,7 @@ public class EventFullPublicDto {
 
     private int views;
 
-    private Map<Emoji, Integer> rates;
+    private Map<Emoji, Long> rates;
 
     private Integer rating;
 }
