+++ ewm-service/src/main/java/ru/practicum/ewmservice/event/dto/EventShortPublicDto.java
@@ -39,7 +39,7 @@ public class EventShortPublicDto {
 
     private int views;
 
-    private Map<Emoji, Integer> rates;
+    private Map<Emoji, Long> rates;
 
     private Integer rating;
 }
